+++ src/maintain/_maintain.js
@@ -3,7 +3,10 @@
 
 import disabled from './disabled';
 import hidden from './hidden';
+import tabFocus from './tab-focus';
+
 export default {
   disabled,
   hidden,
+  tabFocus,
 };